+++ src/Order/LineItemComposer.php
@@ -12,6 +12,7 @@ use ShopgateCartBase;
 use ShopgateExternalCoupon;
 use Shopware\Core\Checkout\Cart\Cart;
 use Shopware\Core\Checkout\Cart\LineItem\LineItem;
+use Shopware\Core\Checkout\Promotion\Cart\Error\PromotionNotEligibleError;
 use Shopware\Core\Checkout\Promotion\Cart\Error\PromotionNotFoundError;
 use Shopware\Core\Content\Product\Cart\ProductNotFoundError;
 use Shopware\Core\Content\Product\Cart\ProductOutOfStockError;
@@ -104,21 +105,31 @@ class LineItemComposer
             switch ($errorClass) {
                 case ProductNotFoundError::class:
                 case ProductOutOfStockError::class:
-                    $missingItem = $sgCart->findItemById($this->productMapping->getIdFromError($error));
-                    if (!$missingItem) {
+                    $id = $this->productMapping->getIdFromError($error);
+                    if (!$missingItem = $sgCart->findItemById($id)) {
+                        $this->logger->debug('Issue locating product: ' . $id);
                         break;
                     }
                     $exportItem = $this->productMapping->mapInvalidProduct($error, $missingItem);
                     $lineItems[$exportItem->getItemNumber()] = $exportItem;
                     break;
                 case PromotionNotFoundError::class:
-                    $missingCoupon = $sgCart->findExternalCoupon($error->getParameters()['code']);
-                    if (!$missingCoupon) {
+                    if (!$missingCoupon = $sgCart->findExternalCoupon($error->getParameters()['code'])) {
+                        $this->logger->debug('Issue locating coupon by code: ' . $error->getParameters()['code']);
                         break;
                     }
                     $missingCoupon->setIsValid(false);
                     $missingCoupon->setNotValidMessage($error->getMessage());
                     $externalCoupons[$error->getParameters()['code']] = $missingCoupon;
+                    break;
+                case PromotionNotEligibleError::class:
+                    if (!$ineligibleCoupon = $sgCart->findExternalCouponByName($error->getParameters()['name'])) {
+                        $this->logger->debug('Issue locating coupon by name: ' . $error->getParameters()['name']);
+                        break;
+                    }
+                    $ineligibleCoupon->setIsValid(false);
+                    $ineligibleCoupon->setNotValidMessage($error->getMessage());
+                    $externalCoupons[] = $ineligibleCoupon;
                     break;
                 default:
                     $this->logger->debug('Unmapped cart errors & notifications');